cleanup + reduce some code dup
